+++ ATK/Special/ConvolutionFilter.h
@@ -33,7 +33,7 @@ namespace ATK
     using Parent::output_sampling_rate;
   protected:
     mutable int64_t split_position;
-    int64_t split_size;
+    int split_size;
     
     FFT<DataType> processor;
